requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken

# uncomment to use huggingface llms
huggingface-hub
InstructorEmbedding
sentence-transformers
langchain_google_genai
langchain_community
langgraph
